feat: start with a task and add libs
