+++ nodemon.json
@@ -1,0 +1,8 @@
+{
+  "watch": ["src"],
+  "ext": "ts, json",
+  "ignore": [
+    "src/**/*.spec.ts"
+  ],
+  "exec": "tsx ./index.ts"
+}
