+++ package.json
@@ -1,21 +1,41 @@
 {
-    "name": "ws_task",
-    "version": "1.0.0",
-    "description": "Template for the WebSocket remote control task",
-    "main": "index.js",
-    "type": "module",
-    "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "author": "Andrei Auchynnikau",
-    "keywords": [],
-    "license": "ISC",
-    "devDependencies": {
-        "nodemon": "^2.0.16"
-    },
-    "dependencies": {
-        "ws": "^8.8.0"
-    }
+  "name"           : "ws_task",
+  "version"        : "1.0.0",
+  "description"    : "Template for the WebSocket remote control task",
+  "main"           : "index.ts",
+  "type"           : "module",
+  "engines"        : {
+    "node": ">=20.0.0"
+  },
+  "scripts"        : {
+    "start"    : "tsx ./index.ts",
+    "start:dev": "nodemon",
+    "test"     : "echo \"Error: no test specified\" && exit 1",
+    "lint"     : "eslint . --ext .ts --fix",
+    "prettier" : "prettier --write \"src/**/*.ts\""
+  },
+  "author"         : "Andrei Auchynnikau",
+  "keywords"       : [],
+  "license"        : "ISC",
+  "devDependencies": {
+    "@types/node"                     : "^20.11.19",
+    "@types/http-server"              : "^0.12.4",
+    "@types/ws"                       : "^8.5.10",
+    "@typescript-eslint/eslint-plugin": "^7.0.1",
+    "@typescript-eslint/parser"       : "^7.0.1",
+    "dotenv"                          : "^16.4.4",
+    "eslint"                          : "^8.56.0",
+    "eslint-config-prettier"          : "^9.1.0",
+    "eslint-plugin-import"            : "^2.29.1",
+    "eslint-plugin-prettier"          : "^5.1.3",
+    "nodemon"                         : "^3.0.3",
+    "prettier"                        : "^3.2.5",
+    "ts-node"                         : "^10.9.2",
+    "ts-node-dev"                     : "^2.0.0",
+    "typescript"                      : "^5.3.3"
+  },
+  "dependencies"   : {
+    "tsx": "^4.7.1",
+    "ws" : "^8.16.0"
+  }
 }
